Ice외형 추가, 변신시 이펙트 추가

+++ Kirby/zzAnimation.h
@@ -39,7 +39,7 @@ namespace zz
 		Texture*					mTex;
 		std::vector<FrameInfo>		mFrames;
 		int							mCurFrame;
-		float						mTime;
+		double						mTime;
 		bool						mFinish;
 		Vector2						mPos;
 	};
